woops, should not use size_t there

+++ HuboCan/src/Hubo2PlusBasicJmc.cpp
@@ -46,7 +46,7 @@ bool Hubo2PlusBasicJmc::decode(const can_frame_t &frame, size_t channel)
             for(size_t i=0; i < joints.size(); ++i)
             {
                 int32_t encoder = 0;
-                for(size_t j=3; j >= 0; --j)
+                for(int j=3; j >= 0; --j)
                 {
                     std::cout << "frame index: " << j + i*4 << std::endl;
                     encoder = (encoder << 8) + frame.data[j + i*4];
